+++ luxonis_train/nodes/heads/base_head.py
@@ -1,10 +1,17 @@
+from pathlib import Path
 from typing import Any
 
+from luxonis_ml.data import DatasetIterator
+from torch import Tensor
+
+from luxonis_train.config.config import PreprocessingConfig
 from luxonis_train.nodes.base_node import (
     BaseNode,
     ForwardInputT,
     ForwardOutputT,
 )
+from luxonis_train.typing import Packet
+from luxonis_train.utils.annotation import default_to_annotate
 
 
 # TODO: We shouldn't skip the head completely
@@ -50,4 +57,26 @@ class BaseHead(BaseNode[ForwardInputT, ForwardOutputT]):
         """
         raise NotImplementedError(
             "get_custom_head_config method must be implemented."
+        )
+
+    def to_annotate(
+        self,
+        head_output: dict[str, Packet[Tensor]],
+        image_paths: list[Path],
+        config_preprocessing: PreprocessingConfig,
+    ) -> DatasetIterator:
+        """Convert head output to a DatasetIterator for dataset annotation. Data should be in standard
+        U{luxonis-ml record format  <https://github.com/luxonis/luxonis-ml/blob/main/luxonis_ml/data/README.md>}.
+
+        @type head_output: dict[str, Packet[Tensor]]
+        @param head_output: Raw outputs from this head.
+        @type image_paths: list[Path]
+        @param image_paths: List of original image file paths.
+        @type config_preprocessing: PreprocessingConfig
+        @param config_preprocessing: Config containing train_image_size, keep_aspect_ratio, etc.
+        @rtype: DatasetIterator
+        @return: Iterator yielding annotation records in luxonis-ml format.
+        """
+        return default_to_annotate(
+            self, head_output, image_paths, config_preprocessing
         )
